Renamed OrbitCore/version.h to OrbitCore/OrbitVersion.h
version.h is a too common name for header. This already lead to name
clashes.
