+++ OrbitGl/App.cpp
@@ -34,6 +34,7 @@
 #include "ModulesDataView.h"
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/Tracing.h"
+#include "OrbitVersion.h"
 #include "OrbitSession.h"
 #include "Params.h"
 #include "Pdb.h"
@@ -46,7 +47,6 @@
 #include "StringManager.h"
 #include "TextRenderer.h"
 #include "Utils.h"
-#include "Version.h"
 
 #if __linux__
 #include <OrbitLinuxTracing/OrbitTracing.h>
