+++ OrbitQt/CrashHandler.cpp
@@ -4,9 +4,9 @@
 
 #include "CrashHandler.h"
 
-#include "../OrbitCore/Version.h"
 #include "Core.h"
 #include "OrbitBase/Logging.h"
+#include "OrbitVersion.h"
 #include "OrbitDbgHelp.h"
 #include "ScopeTimer.h"
 #include "client/settings.h"
