+++ OrbitQt/orbitmainwindow.cpp
@@ -21,11 +21,11 @@
 #include "../OrbitCore/Path.h"
 #include "../OrbitCore/PrintVar.h"
 #include "../OrbitCore/Utils.h"
-#include "../OrbitCore/Version.h"
 #include "../OrbitGl/App.h"
 #include "../OrbitGl/SamplingReport.h"
 #include "../OrbitPlugin/OrbitSDK.h"
 #include "../third_party/concurrentqueue/concurrentqueue.h"
+#include "OrbitVersion.h"
 #include "absl/flags/flag.h"
 #include "absl/strings/match.h"
 #include "absl/strings/str_format.h"
